登陆日志添加

+++ wang-boot-admin/wang-boot-admin-web/src/main/java/com/wpc/system/model/LoginLog.java
@@ -26,7 +26,7 @@ public class LoginLog extends DataEntity<LoginLog> {
     /**
      * 管理员id
      */
-    private Integer userid;
+    private Long userid;
     /**
      * 是否执行成功
      */
@@ -48,11 +48,11 @@ public class LoginLog extends DataEntity<LoginLog> {
         this.logname = logname;
     }
 
-    public Integer getUserid() {
+    public Long getUserid() {
         return userid;
     }
 
-    public void setUserid(Integer userid) {
+    public void setUserid(Long userid) {
         this.userid = userid;
     }
